requirements.txt
discord-py==1.4.1
pillow==7.2.0
requests==2.24.0
pytz  
spotipy